React Router DOM: `matchPath` separation
