+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/imports.kt
@@ -15,8 +15,5 @@ external fun useParams(): Params
 @JsName("useRouteMatch")
 external fun rawUseRouteMatch(options: dynamic): Match?
 
-@JsName("matchPath")
-external fun rawMatchPath(pathName: String, options: dynamic): Match?
-
 @JsName("withRouter")
 external fun <T : Props> rawWithRouter(component: dynamic): ComponentClass<T>
