+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/matchPath.kt
@@ -1,0 +1,10 @@
+@file:JsModule("react-router-dom")
+@file:JsNonModule
+
+package react.router.dom
+
+@JsName("matchPath")
+external fun rawMatchPath(
+    pathName: String,
+    options: dynamic,
+): Match?
