+++ kotlin-react-router-dom/src/main/kotlin/react/router/dom/routingDsl.kt
@@ -1,6 +1,5 @@
 package react.router.dom
 
-import kotlinext.js.jso
 import react.*
 import kotlin.reflect.KClass
 
@@ -153,21 +152,4 @@ fun RBuilder.redirect(
             this.strict = strict
         }
     }
-}
-
-fun matchPath(
-    pathName: String,
-    vararg path: String,
-    exact: Boolean = false,
-    strict: Boolean = false,
-    sensitive: Boolean = false,
-): Match? {
-    val options: RouteMatchOptions = jso {
-        this.path = path
-        this.exact = exact
-        this.strict = strict
-        this.sensitive = sensitive
-    }
-
-    return rawMatchPath(pathName, options)
 }
